Refactor: Remove unused classes and methods
- Deleted EntityField, DomaEntityTask, FlywayTask, FormTask, PatternTemplateTask, BasePackageDetector, EntitySourceAnalyzer, MavenCompileExecutor, MyApplicationFactory, ReplMock, ScaffoldCommandTest, TransportMock, DomaEntityTaskTest, FlywayTaskTest, FormTaskTest, PathResolverMock, PatternTemplateTaskTest, BasePackageDetectorTest, EntitySourceAnalyzerTest, and Product.java.
- Removed associated test resources and templates.
- Updated RoutingMiddleware to handle null responses gracefully.

+++ enkan-core/src/main/java/enkan/collection/OptionMap.java
@@ -62,7 +62,7 @@ public class OptionMap extends HashMap<String, Object> {
         if (value instanceof Number) {
             return ((Number) value).longValue();
         } else {
-            return Integer.parseInt(value.toString());
+            return Long.parseLong(value.toString());
         }
     }
